requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.7
aiofiles>=22.1.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
slowapi>=0.1.9
